requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
openai>=1.0.0
